+++ R/add_initial_values.bvecmodel.R
@@ -26,9 +26,10 @@
 #' draws from the respective prior distributions.
 #' 
 #' In case of a model with time varying parameters (TVP), the initial states are
-#' obtained using the approach specified in argument \code{method}. However, the
-#' initial draws of the error variances of the state equations are always drawn
-#' from their prior distributions. The autocorrelation coefficient \code{rho} of
+#' obtained using the approach specified in argument \code{method}. The precisions
+#' of the state equations start at the means of their gamma priors in case
+#' \code{method = "maxlik"} and are drawn from those priors in case
+#' \code{method = "prior"}. The autocorrelation coefficient \code{rho} of
 #' the state equation of \eqn{\beta} is taken from the prior specification,
 #' whether it is held there or drawn from the prior on it that
 #' \code{coint$rho_min} and \code{coint$rho_max} set up.
@@ -268,7 +269,7 @@ add_initial_values.bvecmodel <- function(object, method = "maxlik", ...){
   }
 
   # Variances of state equations ----
-  object <- .add_initial_values_state_errors(object)
+  object <- .add_initial_values_state_errors(object, method = method)
   
   # Initial values for errors
   object <- .add_initial_values_measurement_errors(object = object,